Fixed: [commit (push) does not work if no devices are connected/mounted](https://github.com/clicon/clixon-controller/issues/5)
Fixed by changing the design so that triggering by commit is removed,
You need to make a rpc_connection_change OPEN after a new device is configured and enabled.
Which means you need to FIRST enable the device and commit, and THEN send a separate RPC when
you enable devices.
And THEREAFTER make a commit push to trigger a service.